Embed API Client: Java
This changeset adds autogenerated API clients in Java for the *Embed
API*, which allows Embedded Menus to be placed on some webpage or other
display. This client is low-level and offers advanced access to
embeddable configuration details.

Clients are provided for both Maven and Gradle build systems, and are
built against API version `v1`.